{{cookiecutter.project_slug}}/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "{{ cookiecutter.project_slug }}"
version = "0.1.0"
description = "{{ cookiecutter.project_description }}"
readme = "README.md"
authors = [{ name = "{{ cookiecutter.full_name }}", email = "{{ cookiecutter.email }}" }]
requires-python = ">={{ cookiecutter.python_version }}"
license = { text = "MIT" }
dependencies = [
  "torch>=2.4.0",
  "torchvision>=0.19.0",
  "torchaudio>=2.4.0",
  "opencv-python>=4.10",
  "pydantic>=2.7",
  "typer>=0.12",
  "rich>=13.7",
  "hydra-core>=1.3",
  "pyyaml>=6.0",
  "einops>=0.8",
  "timm>=1.0",
  "lightning>=2.4",
  "numpy>=1.26",
  "pillow>=10.4"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "mypy>=1.11",
  "ruff>=0.5",
  "pre-commit>=3.7",
  "ipykernel>=6.29"
]
deploy = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "onnx>=1.16",
  "onnxruntime>=1.18",
  "boto3>=1.34",
  "sanic>=23.12"
]
ml = [
  "mlflow>=2.15",
  "pandas>=2.2",
  "scikit-learn>=1.5"
]

[project.scripts]
train_model = "scripts.train:app"
export_model = "scripts.export:app"
serve_model = "scripts.serve:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.uv]
default-groups = ["dev"]
resolver.map-prebuilt = true

[tool.ruff]
line-length = 100
target-version = "py{{ cookiecutter.python_version.replace('.', '') }}"
lint.select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.mypy]
python_version = "{{ cookiecutter.python_version }}"
strict = false
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-q", "--disable-warnings"]

[tool.coverage.run]
source = ["src/{{ cookiecutter.python_package }}"]
branch = true

